+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.5.0] - 2019-06-12
+
+- [ADDED] `huh:migration:news_tags` command
+
 ## [0.4.1] - 2019-06-06
 
 - fixed id issue in news categories migration